Renaming of implementation files to *.impl.ts
