+++ src/impl/AtomicBoolean.impl.ts
