+++ src/impl/AtomicInteger.impl.ts
