+++ src/impl/Repository.impl.ts
@@ -126,17 +126,17 @@ class RepositoryImpl implements Repository {
     readonly #requiredContracts: Set<Contract<unknown>> = new Set<Contract<unknown>>();
 }
 
-import { RequiredType } from "../api/Types";
-import { Contract } from "../api/Contract";
-import { Contracts } from "../api/Contracts";
-import { Promisor, PromisorType, typeToPromisor } from "../api/Promisor";
-import { Repository } from "../api/Repository";
 import { AUTO_CLOSE_NONE, AutoClose, inlineAutoClose } from "../api/AutoClose";
 import { BindStrategy, resolveBindStrategy } from "../api/BindStrategy";
 import { contractCheck } from "../api/Checks";
+import { Contract } from "../api/Contract";
 import { ContractException } from "../api/ContractException";
+import { Contracts } from "../api/Contracts";
+import { Promisor, PromisorType, typeToPromisor } from "../api/Promisor";
+import { Repository } from "../api/Repository";
+import { RequiredType } from "../api/Types";
 
-import { StorageImpl } from "./StorageImpl";
-import { IdempotentImpl } from "./IndempotentImpl";
 import { OptionalType } from "../api/Types";
+import { IdempotentImpl } from "./Indempotent.impl";
+import { StorageImpl } from "./Storage.impl";
